fix: Merge Order and UploadDoc
