+++ src/main/java/com/greenatom/controller/OrderController.java
@@ -5,6 +5,7 @@ import com.greenatom.controller.api.OrderApi;
 import com.greenatom.domain.dto.order.GenerateOrderRequestDTO;
 import com.greenatom.domain.dto.order.OrderRequestDTO;
 import com.greenatom.domain.dto.order.OrderResponseDTO;
+import com.greenatom.domain.dto.order.UploadDocumentRequestDTO;
 import com.greenatom.service.OrderService;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -12,6 +13,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -82,5 +84,11 @@ public class OrderController implements OrderApi {
             produces = {"application/json"})
     public void deleteOrder(@PathVariable Long id) {
         orderService.deleteOrder(id);
+    }
+
+    @PostMapping(value = "/upload", consumes = "multipart/form-data", produces = {"application/json"})
+    public void uploadFile(@RequestParam("file") MultipartFile file, @RequestParam("id") Long id) {
+        UploadDocumentRequestDTO uploadDocumentRequestDTO = new UploadDocumentRequestDTO(file, id);
+        orderService.upload(uploadDocumentRequestDTO);
     }
 }
