+++ src/main/java/com/greenatom/controller/api/OrderApi.java
@@ -3,6 +3,7 @@ package com.greenatom.controller.api;
 import com.greenatom.domain.dto.order.GenerateOrderRequestDTO;
 import com.greenatom.domain.dto.order.OrderRequestDTO;
 import com.greenatom.domain.dto.order.OrderResponseDTO;
+import com.greenatom.domain.dto.security.JwtResponse;
 import com.greenatom.utils.exception.message.OrderErrorMessage;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -11,7 +12,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springdoc.api.ErrorMessage;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -244,4 +247,35 @@ public interface OrderApi {
             Long id
     );
 
+    @ApiResponses(value = {
+        @ApiResponse(
+            responseCode = "200",
+            description = "Документ успешно загружен",
+            content = {
+                    @Content(
+                            mediaType = "application/json",
+                            schema = @Schema(implementation = JwtResponse.class)
+                    )
+            }
+        ),
+    @ApiResponse(
+            responseCode = "400",
+            description = "Неверный запрос или невозможно загрузить документ",
+            content = {
+                    @Content(
+                            mediaType = "application/json",
+                            schema = @Schema(implementation = ErrorMessage.class)
+                    )
+            }
+    )
+            })
+    @Operation(
+            summary = "Загрузить документ"
+    )
+    void uploadFile(
+            @Parameter(description = "Выберите подписанный документ (любой формат)")
+            MultipartFile file,
+            @Parameter(description = "Введите orderID, к которому хотите прикрепить документ", example = "1")
+            Long id
+    );
 }
