--- src/main/java/com/greenatom/service/UploadDocService.java
@@ -1,15 +1,0 @@
-package com.greenatom.service;
-
-import com.greenatom.domain.dto.order.UploadDocumentRequestDTO;
-import org.springframework.transaction.annotation.Transactional;
-
-public interface UploadDocService {
-
-    @Transactional
-    void upload(UploadDocumentRequestDTO uploadDocumentRequestDTO);
-
-    @Transactional
-    void updateStatus(UploadDocumentRequestDTO uploadDocumentRequestDTO);
-    @Transactional
-    void updatePath(UploadDocumentRequestDTO uploadDocumentRequestDTO);
-}
