+++ src/main/java/com/greenatom/service/impl/OrderServiceImpl.java
@@ -4,6 +4,7 @@ import com.greenatom.domain.dto.item.OrderItemRequestDTO;
 import com.greenatom.domain.dto.order.GenerateOrderRequestDTO;
 import com.greenatom.domain.dto.order.OrderRequestDTO;
 import com.greenatom.domain.dto.order.OrderResponseDTO;
+import com.greenatom.domain.dto.order.UploadDocumentRequestDTO;
 import com.greenatom.domain.entity.*;
 import com.greenatom.domain.enums.OrderStatus;
 import com.greenatom.domain.mapper.OrderMapper;
@@ -19,6 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.poi.xwpf.usermodel.XWPFDocument;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.*;
 import java.util.List;
@@ -108,6 +110,7 @@ public class OrderServiceImpl implements OrderService {
         return orderMapper.toDto(order);
     }
 
+
     private Order createDraftOrder(OrderRequestDTO orderRequestDTO) {
         Order order = new Order();
         Client client = clientRepository
@@ -193,4 +196,65 @@ public class OrderServiceImpl implements OrderService {
             log.error("Couldn't convert file");
         }
     }
+
+    //Загрузка подписанного документа
+    @Override
+    public void upload(UploadDocumentRequestDTO uploadDocumentRequestDTO) {
+        MultipartFile file = uploadDocumentRequestDTO.getFile();
+        if (!file.isEmpty()) {
+            try {
+                String projectRoot = System.getProperty("user.dir");
+                String uploadDir = projectRoot + "/Documents/UploadDoc";
+                String fileName = cleanFileName(file.getOriginalFilename());
+
+                File uploadPath = new File(uploadDir);
+                if (!uploadPath.exists()) {
+                    uploadPath.mkdirs();
+                }
+
+                File targetFile = new File(uploadPath, fileName);
+                try (FileOutputStream fos = new FileOutputStream(targetFile)) {
+                    fos.write(file.getBytes());
+                }
+                uploadDocumentRequestDTO.setLinkToFolder(targetFile.getAbsolutePath());
+                log.info("The file has been successfully uploaded. File name: " + fileName + ", Path: "
+                        + targetFile.getAbsolutePath());
+            } catch (IOException e) {
+                log.error("Error uploading file: " + e.getMessage());
+            }
+        } else {
+            log.error("File is empty, download failed.");
+        }
+        updateStatus(uploadDocumentRequestDTO);
+    }
+
+    //Обновляем статус в заявке на SIGNED_BY_CLIENT
+    private void updateStatus(UploadDocumentRequestDTO uploadDocumentRequestDTO) {
+        Order order = orderRepository
+                .findById(uploadDocumentRequestDTO.getId())
+                .orElseThrow(OrderException.CODE.NO_SUCH_ORDER::get);
+        if (order.getOrderStatus().equals(OrderStatus.SIGNED_BY_EMPLOYEE)) {
+
+            order.setOrderStatus(OrderStatus.SIGNED_BY_CLIENT);
+        } else {
+            throw OrderException.CODE.CANNOT_ASSIGN_ORDER.get();
+        }
+        updatePath(uploadDocumentRequestDTO);
+    }
+
+    private void updatePath(UploadDocumentRequestDTO uploadDocumentRequestDTO) {
+        log.debug("Order to update link_to_folder : {}", uploadDocumentRequestDTO);
+
+        Long orderId = uploadDocumentRequestDTO.getId();
+        String linkToFolder = uploadDocumentRequestDTO.getLinkToFolder();
+
+        orderRepository.updateLinkToFolder(orderId, linkToFolder);
+    }
+
+    private String cleanFileName(String fileName) {
+        return fileName.replaceAll("[^a-zA-Z0-9_-]", "");
+    }
+
+
+
 }
